% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSampleData.R
\name{GetSampleData}
\alias{GetSampleData}
\title{Creating Sample Data for Plotting}
\usage{
GetSampleData()
}
\value{
A data.frame with the appropriate shape for the plotting function ReportPlot.
}
\description{
Creating Sample Data for Plotting
}
\examples{
data <- GetSampleData()
}
