% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportPlot.R
\name{ReportPlot}
\alias{ReportPlot}
\title{Generating the Plot}
\usage{
ReportPlot(x, main = "Productname", ylab = "cpm (Moving Average)",
  connection = "spline",
  legend_line = "Compliants per Million sold; \\nMoving Average on 3 Month base",
  legend_bar = "Quantity Sold")
}
\arguments{
\item{x}{A data.frame with the shape of: See GetSampleData}

\item{main}{Title of the plot.}

\item{ylab}{Label for the y-Axis.}

\item{connection}{Can be either spline or linear}

\item{legend_line}{Legend for curve/line in the plot}

\item{legend_bar}{Legeng for bars in the plot}
}
\value{
Returns the Plot.
}
\description{
The function was developed for a shiny report of a medical device manufacturer. \href{https://www.data-science-archtect.de}{More about the project and developer: Data-Science-Architect}.
}
\examples{
data <- GetSampleData()
ReportPlot(data)
}
