% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleData.R
\docType{data}
\name{SampleData}
\alias{SampleData}
\title{Sample Data for ReportPlot}
\format{A data frame with 48 rows and 4 variables:
\describe{
  \item{year}{year of the data}
  \item{month}{month of the data}
  \item{sold}{quantity / sold units}
  \item{cpm}{complaints per million}
}}
\usage{
data(SampleData)
}
\description{
The data has the appropriate structure for the ReportPlot and served during development. Now it clearifies the internals and usage.
}
\examples{
data(SampleData)
}
\keyword{datasets}
